#75 some refactoring to modularize test code
